r/legacy_vmss: moving the compute state upgrader into that package
