+++ azurerm/internal/services/compute/migration/legacy_vmss.go
